requirements.txt
telethon==1.36.0
pytesseract
Pillow
sympy